add some site commaner and some rpc
添加一些站点命令和rpc命令
