+++ lib/cli/index.js
@@ -49,6 +49,8 @@ log.i18nInit().then(() => {
   // 隐藏输出内容
   commander.option('-s, --silent', log.t('cli.output.description.silent'), false)
 }).then(() => {
+  // 引入rpc模块
+  require('./rpc')
   // 引入site管理模块
   require('./siteManage')
   // 引入server管理模块
