+++ lib/cli/serverManage/index.js
@@ -10,3 +10,33 @@ const commander = require('commander')
 //log.show('开发中')
 
 //console.log(config)
+
+/**
+ * 开启服务进程
+ */
+commander
+.command('start')
+.description(log.t('cli.help.command.start'))
+.action(function start () { commander.rpcBySiteIdName('start') })
+/**
+* 重启服务进程
+*/
+commander
+.command('restart')
+.description('Restart ddv server or restart site')
+.description(log.t('cli.help.command.restart'))
+.action(function restart () { commander.rpcBySiteIdName('restart') })
+/**
+* 重启服务进程
+*/
+commander
+.command('reload')
+.description(log.t('cli.help.command.reload'))
+.action(function reload () { commander.rpcBySiteIdName('reload') })
+/**
+* 停止服务进程
+*/
+commander
+.command('stop')
+.description(log.t('cli.help.command.stop'))
+.action(function stop () { commander.rpcBySiteIdName('stop') })
