+++ lib/cli/siteManage/index.js
@@ -7,7 +7,82 @@ const log = require(ddvPath('lib/log'))
 // commander 模块
 const commander = require('commander')
 
-//log.show('开发中')
+/**
+ * 添加站点
+ */
+commander
+.command('add <file|path>')
+.option('-n, --name <name>', 'set a <name> for site')
+.description(log.t('cli.help.command.add'))
+.action(function (path) {
+  c.__api('add', c._cmdGetPathName(path, c.name), function callback (e, res) {
+    if (e) {
+      log.tip('FAIL', 'CLI_COMMAND_ADD_STIE_FAIL')
+      console.error(colors.grey(e.message))
+      console.error('')
+      console.error(e)
+      c.__disconnect()
+    } else {
+      log.tip(' OK ', 'CLI_COMMAND_ADD_STIE_SUCCESS')
+      // 显示列表
+      c._cmdLists()
+    }
+  })
+  path = undefined
+})
 
-//console.log(config)
-//console.log(config.DDV_OUT_LOG_FILE_PATH)
+/**
+ * 开启服务进程
+ */
+commander
+.command('start [name|siteId]')
+.option('-n, --name <names>', '<names> eg: -n app1,app2')
+.option('-i, --siteId <siteIds>', '<siteIds> eg: -s 1,2', '')
+.description(log.t('cli.help.command.start_site'))
+.action(function start () { commander.rpcBySiteIdName('start') })
+/**
+* 重启服务进程
+*/
+commander
+.command('restart [name|siteId]')
+.option('-n, --name <names>', '<names> eg: -n app1,app2')
+.option('-i, --siteId <siteIds>', '<siteIds> eg: -s 1,2', '')
+.description('Restart ddv server or restart site')
+.description(log.t('cli.help.command.restart_site'))
+.action(function restart () { commander.rpcBySiteIdName('restart') })
+/**
+* 停止服务进程
+*/
+commander
+.command('stop [name|siteId]')
+.option('-n, --name <names>', '<names> eg: -n app1,app2')
+.option('-i, --siteId <siteIds>', '<siteIds> eg: -s 1,2', '')
+.description(log.t('cli.help.command.stop_site'))
+.action(function stop () { commander.rpcBySiteIdName('stop') })
+/**
+ * 移除
+*/
+commander
+.command('remove [name|siteId]')
+.option('-n, --name <names>', '<names> eg: -n app1,app2', '')
+.option('-i, --siteId <siteIds>', '<siteIds> eg: -s 1,2', -1)
+.description(log.t('cli.help.command.remove'))
+.action(function remove () { commander.rpcBySiteIdName('remove') })
+/**
+ * 移除 remove的别名
+*/
+commander
+.command('delete [name|siteId]')
+.option('-n, --name <names>', '<names> eg: -n app1,app2')
+.option('-i, --siteId <siteIds>', '<siteIds> eg: -s 1,2', '')
+.description(log.t('cli.help.command.remove_alias'))
+.action(function remove () { commander.rpcBySiteIdName('remove') })
+/**
+ * 移除 remove的别名
+*/
+commander
+.command('del [name|siteId]')
+.option('-n, --name <names>', '<names> eg: -n app1,app2')
+.option('-i, --siteId <siteIds>', '<siteIds> eg: -s 1,2', '')
+.description(log.t('cli.help.command.remove_alias'))
+.action(function remove () { commander.rpcBySiteIdName('remove') })
