+++ lib/locales/en_US/cli/index.js
@@ -32,7 +32,17 @@ module.exports = {
       remove_site_help: 'Remove site help'
     },
     command: {
-      tail: '\'site_name\' is the name of the site, The function of this command is to monitor the log changes, and real-time output'
+      tail: '\'site_name\' is the name of the site, The function of this command is to monitor the log changes, and real-time output',
+      add: 'Add an site + set a name',
+      remove: 'Remove an site',
+      remove_alias: 'Remove an site(remove alias)',
+      start_site: 'Start site',
+      restart_site: 'Restart site',
+      stop_site: 'Stop site',
+      start: 'Start ddv server',
+      restart: 'Restart ddv server',
+      reload: 'Reload the ddv server site changes',
+      stop: 'Stop ddv server'
     }
   },
   command: {
