+++ lib/locales/zh_CN/cli/index.js
@@ -32,7 +32,17 @@ module.exports = {
       remove_site_help: '删除站点帮助'
     },
     command: {
-      tail: '\'site_name\'是网站名称，此命令的功能是监视日志更改，并实时输出'
+      tail: '\'site_name\'是网站名称，此命令的功能是监视日志更改，并实时输出',
+      add: '添加一个站点，并且设置站点名字',
+      remove: '删除一个站点',
+      remove_alias: '删除一个站点(remove 别名)',
+      start_site: '启动指定站点',
+      restart_site: '重启指定站点',
+      stop_site: '停止指定站点',
+      start: '启动DDV服务',
+      restart: '重启DDV服务',
+      reload: '重新加载DDV服务器网站的更改',
+      stop: '停止DDV服务'
     }
   },
   command: {
